new changes made

+++ routes/web.php
@@ -23,25 +23,26 @@ Auth::routes();
     ]);
     Route::get('/payment/callback', 'PaymentController@handleGatewayCallback');
 
-   
+
 
             Route::get('/payment/{plan}/plan','PaymentController@paymentplan');
             Route::put('/payment/{plan}/update','PaymentController@updatePlan');
-            Route::get('/card/payment','PaymentController@paymentForm');
-            Route::get('/dashboard','DashboardController@dashboard');
+            Route::get('/card/payment','PaymentController@paymentForm')->middleware('auth');
+
+            Route::get('/dashboard','DashboardController@dashboard')->middleware('auth');
 
             // Route::get('chart-line', 'ChartController@chartLine');
             Route::get('chart-line-ajax', 'DashboardController@chartLineAjax');
 
 
-            Route::get('/import/','StateController@importState');
-            Route::post('/import-data/','StateController@importStateData');
-            Route::get('/downloadExcel/{type}','StateController@exportStateData');
+            Route::get('/import/','StateController@importState')->middleware(['auth','admin']);
+            Route::post('/import-data/','StateController@importStateData')->middleware(['auth','admin']);
+            Route::get('/downloadExcel/{type}','StateController@exportStateData')->middleware(['auth','admin']);
 
 
             Route::group(['prefix'=> 'patients'] , function(){
 
-                Route::get('dashboard','PatientController@dashboard');
+                Route::get('dashboard','PatientController@dashboard')->middleware('auth');
                 Route::get('/register', 'PatientController@create');
                 Route::Post('/store','PatientController@store');
                 Route::get('/store/step-two','PatientController@stepTwo');
@@ -54,13 +55,20 @@ Auth::routes();
 
                 Route::get('/preview','PatientController@preview');
                 Route::get('/store/data','PatientController@storeArrayData');
-                Route::get('/all','PatientController@allPatient');
-                Route::get('/{patient}/show','PatientController@show');
+                Route::get('/all','PatientController@allPatient')->middleware(['auth','admin']);
+                Route::get('/{patient}/show','PatientController@show')->middleware(['auth']);
 
                 //store patient complaints/request
-                Route::get('/complaints' ,'ComplaintController@request');
-                Route::post('/complaints/store','ComplaintController@store');
-                Route::get('/{complaint}/complaints/history','ComplaintController@complaintHistory');
+                Route::get('/complaints' ,'ComplaintController@request')->middleware(['auth']);
+                Route::post('/complaints/store','ComplaintController@store')->middleware(['auth']);
+                Route::get('/{complaint}/complaints/history','ComplaintController@complaintHistory')->middleware(['auth']);
+
+                //feedback need to implement new feature before final push
+                //create a new controller feed back 
+                //fetch prescription id and useri id 
+                //create a tabel that relates to feed back 
+                
+                Route::post('/{prescription}/feedback' ,'FeedbackController@feedback');
 
             });
 
@@ -69,8 +77,9 @@ Auth::routes();
 
             Route::group(['prefix'=> 'complaints'] , function(){
 
-                Route::get('/all/complaints','ComplaintController@allComplaints');
-                Route::get('/{complaint}/show/','ComplaintController@viewComplaints');
+                Route::get('/all/complaints','ComplaintController@allComplaints')->middleware(['auth','admin']);
+                Route::get('/{complaint}/show/','ComplaintController@viewComplaints')->middleware(['auth']);
+                Route::get('/{complaint}/prescription','ComplaintController@findPrescriptionForComplaint')->middleware(['auth']);
 
             });
 
@@ -81,18 +90,21 @@ Auth::routes();
                 Route::get('/{prescription}/view','PrescriptionController@viewPrescription');
 
                 //all prescription
-                Route::get('/all','PrescriptionController@allPrescription')->middleware('admin');
+                Route::get('/all','PrescriptionController@allPrescription')->middleware(['auth','admin']);
+
+                //user prescriptions
+                Route::get('/{prescription}/user','PrescriptionController@userPrescriptions')->middleware('auth');
 
             });
 
         Route::group(['prefix'=> 'subscriptions'] , function(){
 
-                Route::get('all','SubscriptionController@allSubscriptions')->middleware('admin');
-                Route::get('add','SubscriptionController@create')->middleware('admin');
-                Route::post('store','SubscriptionController@store')->middleware('admin');
-                Route::get('/{subscription}/edit','SubscriptionController@edit')->middleware('admin');
-                Route::post('/{subscription}/update','SubscriptionController@update')->middleware('admin');
-                Route::delete('/{subscription}/delete','SubscriptionController@delete')->middleware('admin');
+                Route::get('all','SubscriptionController@allSubscriptions')->middleware('auth');
+                Route::get('add','SubscriptionController@create')->middleware(['auth','admin']);
+                Route::post('store','SubscriptionController@store')->middleware(['auth','admin']);
+                Route::get('/{subscription}/edit','SubscriptionController@edit')->middleware(['auth','admin']);
+                Route::post('/{subscription}/update','SubscriptionController@update')->middleware(['auth','admin']);
+                Route::delete('/{subscription}/delete','SubscriptionController@delete')->middleware(['auth','admin']);
                 Route::get('/plan','SubscriptionController@selectPlan')->name('subscriptions');
                 Route::post('/{plan}/subscribed','SubscriptionController@subscribed');
 
@@ -102,18 +114,18 @@ Auth::routes();
 
                 Route::group(['prefix'=> 'doctors'] , function(){
 
-                    Route::get('all','DoctorController@allDoctors');
-                    Route::get('add','DoctorController@addDoctors')->middleware('admin');
-                    Route::post('store','DoctorController@store')->middleware('admin');
-                    Route::get('/{doctor}/edit','DoctorController@edit');
-                    Route::post('/{doctor}/update','DoctorController@update');
+                    Route::get('all','DoctorController@allDoctors')->middleware(['auth','admin']);
+                    Route::get('add','DoctorController@addDoctors')->middleware(['auth','admin']);
+                    Route::post('store','DoctorController@store')->middleware(['auth','admin']);
+                    Route::get('/{doctor}/edit','DoctorController@edit')->middleware(['auth','admin']);
+                    Route::post('/{doctor}/update','DoctorController@update')->middleware(['auth','admin']);
                     // Route::delete('/{subscription}/delete','SubscriptionController@delete');
 
                 });
            });
  
             Route::group(['prefix'=> 'settings'] , function(){
-                Route::get('/{setting}/profile','SettingsController@profile');
+                Route::get('/{setting}/profile','SettingsController@profile')->middleware(['auth']);
             });
